// src/generated/client/models/DailyEntry.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Timebank
 * Timebank API documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * Person's daily entry
 * @export
 * @interface DailyEntry
 */
export interface DailyEntry {
    /**
     * Person ID in Forecast
     * @type {number}
     * @memberof DailyEntry
     */
    person: number;
    /**
     * Amount of internal time per daily entry
     * @type {number}
     * @memberof DailyEntry
     */
    internalTime: number;
    /**
     * Amount of billable project time per daily entry
     * @type {number}
     * @memberof DailyEntry
     */
    billableProjectTime: number;
    /**
     * Amount of non billable project time per daily entry
     * @type {number}
     * @memberof DailyEntry
     */
    nonBillableProjectTime: number;
    /**
     * Amount of time logged per daily entry
     * @type {number}
     * @memberof DailyEntry
     */
    logged: number;
    /**
     * Amount of project time logged per daily entry
     * @type {number}
     * @memberof DailyEntry
     */
    loggedProjectTime: number;
    /**
     * Amount of time expected per daily entry
     * @type {number}
     * @memberof DailyEntry
     */
    expected: number;
    /**
     * Total amount of worktime per daily entry
     * @type {number}
     * @memberof DailyEntry
     */
    balance: number;
    /**
     * Date of daily entry
     * @type {Date}
     * @memberof DailyEntry
     */
    date: Date;
    /**
     * Is daily entry vacation or not
     * @type {boolean}
     * @memberof DailyEntry
     */
    isVacation: boolean;
}

/**
 * Check if a given object implements the DailyEntry interface.
 */
export function instanceOfDailyEntry(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "person" in value;
    isInstance = isInstance && "internalTime" in value;
    isInstance = isInstance && "billableProjectTime" in value;
    isInstance = isInstance && "nonBillableProjectTime" in value;
    isInstance = isInstance && "logged" in value;
    isInstance = isInstance && "loggedProjectTime" in value;
    isInstance = isInstance && "expected" in value;
    isInstance = isInstance && "balance" in value;
    isInstance = isInstance && "date" in value;
    isInstance = isInstance && "isVacation" in value;

    return isInstance;
}

export function DailyEntryFromJSON(json: any): DailyEntry {
    return DailyEntryFromJSONTyped(json, false);
}

export function DailyEntryFromJSONTyped(json: any, ignoreDiscriminator: boolean): DailyEntry {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'person': json['person'],
        'internalTime': json['internalTime'],
        'billableProjectTime': json['billableProjectTime'],
        'nonBillableProjectTime': json['nonBillableProjectTime'],
        'logged': json['logged'],
        'loggedProjectTime': json['loggedProjectTime'],
        'expected': json['expected'],
        'balance': json['balance'],
        'date': (new Date(json['date'])),
        'isVacation': json['isVacation'],
    };
}

export function DailyEntryToJSON(value?: DailyEntry | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'person': value.person,
        'internalTime': value.internalTime,
        'billableProjectTime': value.billableProjectTime,
        'nonBillableProjectTime': value.nonBillableProjectTime,
        'logged': value.logged,
        'loggedProjectTime': value.loggedProjectTime,
        'expected': value.expected,
        'balance': value.balance,
        'date': (value.date.toISOString().substring(0,10)),
        'isVacation': value.isVacation,
    };
}