// src/generated/client/models/Timespan.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Timebank
 * Timebank API documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Enum for timespan
 * @export
 */
export const Timespan = {
    ALL_TIME: 'ALL_TIME',
    YEAR: 'YEAR',
    MONTH: 'MONTH',
    WEEK: 'WEEK'
} as const;
export type Timespan = typeof Timespan[keyof typeof Timespan];


export function TimespanFromJSON(json: any): Timespan {
    return TimespanFromJSONTyped(json, false);
}

export function TimespanFromJSONTyped(json: any, ignoreDiscriminator: boolean): Timespan {
    return json as Timespan;
}

export function TimespanToJSON(value?: Timespan | null): any {
    return value as any;
}